requirements.txt
appdirs==1.4.3
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.0
dnspython==1.16.0
email-validator==1.1.0
eventlet==0.25.2
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
googlemaps==4.3.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
monotonic==1.5
Naked==0.1.31
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycryptodomex==3.9.7
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
python-bsonjs==0.2.0
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
toml==0.10.0
tox==3.15.0
urllib3==1.25.9
virtualenv==20.0.20
waitress==1.4.3
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
